# Dumps/Source Code/8 - Development_website/master/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Turtle WoW API",
        "version": "0.1"
    },
    "paths": {
        "/api/characters/population": {
            "get": {
                "tags": [
                    "Character"
                ],
                "summary": "Get population statistics",
                "operationId": "getPopulationStatistics",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "week_name": {
                                                "type": "string"
                                            },
                                            "year": {
                                                "type": "integer"
                                            },
                                            "week": {
                                                "type": "integer"
                                            },
                                            "m_a": {
                                                "type": "integer"
                                            },
                                            "m_h": {
                                                "type": "integer"
                                            },
                                            "date_nice": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
...
}